+++ inst/doc/Example-National-Ambulatory-Medical-Care-Survey-NAMCS-tables.R
@@ -1,17 +1,16 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE
   # , comment = "#>"
 )
 
-## ---- results=FALSE, message=FALSE, warning=FALSE-----------------------------
+## ----results=FALSE, message=FALSE, warning=FALSE------------------------------
 library(surveytable)
 
-## -----------------------------------------------------------------------------
-mysurvey = namcs2019sv
-set_survey(mysurvey)
+## ----results='asis'-----------------------------------------------------------
+set_survey(namcs2019sv)
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 total()
 tab("MDDO", "SPECCAT", "MSA")
 
@@ -22,23 +21,20 @@ names(uspop2019)
 ## -----------------------------------------------------------------------------
 uspop2019$total
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 total_rate(uspop2019$total)
-
-## -----------------------------------------------------------------------------
-levels(mysurvey$variables$MSA)
 
 ## -----------------------------------------------------------------------------
 uspop2019$MSA
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_rate("MSA", uspop2019$MSA)
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_rate("MDDO", uspop2019$total)
 tab_rate("SPECCAT", uspop2019$total)
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 var_list("age")
 
 ## -----------------------------------------------------------------------------
@@ -46,11 +42,11 @@ var_cut("Age group", "AGE"
         , c(-Inf, 0, 4, 14, 64, Inf)
         , c("Under 1", "1-4", "5-14", "15-64", "65 and over") )
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab("AGER", "Age group", "SEX")
 tab_cross("AGER", "SEX")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_rate("AGER", uspop2019$AGER)
 tab_rate("Age group", uspop2019$`Age group`)
 tab_rate("SEX", uspop2019$SEX)
@@ -58,34 +54,39 @@ tab_rate("SEX", uspop2019$SEX)
 ## -----------------------------------------------------------------------------
 uspop2019$`AGER x SEX`
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_subset_rate("AGER", "SEX", uspop2019$`AGER x SEX`)
 
-## -----------------------------------------------------------------------------
-mysurvey$variables = within(mysurvey$variables, {
-  `Medicare and Medicaid` = PAYMCARE & PAYMCAID
-  `Unknown or blank` = PAYDK | (NOPAY == "No categories marked")
-  nopwu = !( PAYPRIV | PAYMCARE | PAYMCAID | PAYWKCMP
-                | PAYOTH | PAYDK)
-  `Self-pay (real)` = PAYSELF & nopwu
-  `No charge (real)` = PAYNOCHG & nopwu
-  `No insurance` = `Self-pay (real)` | `No charge (real)`
-})
-set_survey(mysurvey)
+## ----results='asis'-----------------------------------------------------------
+#
+var_all("Medicare and Medicaid", c("PAYMCARE", "PAYMCAID"))
 
-## ---- results='asis'----------------------------------------------------------
+#
+var_any("Payment used", c("PAYPRIV", "PAYMCARE", "PAYMCAID"
+  , "PAYWKCMP", "PAYOTH", "PAYDK"))
+var_not("No other payment used", "Payment used")
+
+var_all("Self-pay", c("PAYSELF", "No other payment used"))
+var_all("No charge", c("PAYNOCHG", "No other payment used"))
+var_any("No insurance", c("Self-pay", "No charge"))
+
+#
+var_case("No pay", "NOPAY", "No categories marked")
+var_any("Unknown or blank", c("PAYDK", "No pay"))
+
+##
 tab("PAYPRIV", "PAYMCARE", "PAYMCAID", "Medicare and Medicaid"
-    , "No insurance", "Self-pay (real)", "No charge (real)"
-    , "PAYWKCMP", "PAYOTH", "Unknown or blank")
+  , "No insurance", "Self-pay", "No charge"
+  , "PAYWKCMP", "PAYOTH", "Unknown or blank")
 
 ## -----------------------------------------------------------------------------
 var_collapse("PRIMCARE", "Unknown if PCP", c("Unknown", "Blank"))
 var_collapse("REFER", "Unknown if referred", c("Unknown", "Blank"))
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab("PRIMCARE", "REFER", "SENBEFOR")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_subset("PRIMCARE", "SENBEFOR")
 tab_subset("REFER", "SENBEFOR")
 
@@ -95,16 +96,16 @@ var_cut("Age group", "AGE"
         , c("Under 1", "1-4", "5-14", "15-64", "65 and over") )
 var_cross("Age x Sex", "AGER", "SEX")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab("MAJOR")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 tab_subset("AGER", "MAJOR", "Preventive care")
 tab_subset("Age group", "MAJOR", "Preventive care")
 tab_subset("SEX", "MAJOR", "Preventive care")
 tab_subset("Age x Sex", "MAJOR", "Preventive care")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 for (vr in c("AGER", "Age group", "SEX", "Age x Sex")) {
 	print( tab_subset(vr, "MAJOR", "Preventive care") )
 }
@@ -116,14 +117,14 @@ suppressMessages( set_output(csv = tmp_file) )
 for (vr in c("AGER", "Age group", "SEX", "Age x Sex")) {
 	var_cross("tmp", "MAJOR", vr)
 	for (lvl in levels(surveytable:::env$survey$variables[,vr])) {
-		tab_subset("SPECCAT", "tmp", paste0("Preventive care : ", lvl))
+		tab_subset("SPECCAT", "tmp", paste0("Preventive care: ", lvl))
 	}
 }
 set_output(csv = "")
 
-## ---- results='asis'----------------------------------------------------------
+## ----results='asis'-----------------------------------------------------------
 vr = "AGER"
 var_cross("tmp", "MAJOR", vr)
 lvl = levels(surveytable:::env$survey$variables[,vr])[1]
-tab_subset("SPECCAT", "tmp", paste0("Preventive care : ", lvl))
+tab_subset("SPECCAT", "tmp", paste0("Preventive care: ", lvl))