+++ NEWS.md
@@ -1,3 +1,17 @@
+# surveytable 0.9.4
+
+* Optionally adjust p-values for multiple comparisons (`p_adjust` argument)
+
+# surveytable 0.9.3
+
+* `codebook()`
+* `set_survey()`:
+  * Improved output.
+  * Allows an unweighted survey as a `data.frame`.
+  * Can set certain options using an argument.
+* Tabulation functions show the number of observations.
+* LaTeX printing.
+
 # surveytable 0.9.2
 
 * Addressed CRAN comments.